backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-Limiter==3.8.0
